Add helper.cpp
